feat: revive monorepo
This is work in progress.
The proposed monorepo change would allow us ship both CJS, ESM and DTS
files split in different proxy packages for better tree-shacking
purposes.

As well as it "fixes" (and ignores) some of the issues within the `core`
package, and yet needs more type-safety fixes on `react` package.