feat: working `init` command
Turns out I forgot to install some dependencies to the `cli` package
and it made ESBuild to go crazy.
I also had to implement a hack to replace the  in the  field
with a callback since it is a required plugin as of now.
It will not be required in the future versions of the cli.