feat: `core` refactor
Problem: The current state of the core is not reliable for use.
Some of the parameters of the functions are simply ignored (faced that
with `apiRequest` previously), some are counter-intuitive and use
repetitive naming. Matter fact, ~30% code of the classes is its
initialization, and sometimes the actual business logic are just two or
even a single function.

Solution:
- Move away from writing custom classes on top of the other wherever
  possible and rely on the `@abstract-money/cli` generated modules as
  much as possible.
- Provide intuitive and declarative naming to the methods to be used.
- De-classify classes that should not have been classes and reduce
  complexity.
- To eliminate confusion between `AnsAsset` and `Asset` (which has an
  `amount` field), introduce `Token` type.

  The list will be populated as the refactor goes.