feat: clients, separation of concerns
I figured out that we actually have 5 clients:
- base client – knows only the `apiUrl`
- public client - knows the `apiUrl` and `CosmWasmClient`
- wallet client - knows the `apiUrl` and `SigningCosmWasmClient`
- account-public client - knows it's accountId, `apiUrl` and `CosmWasmClient`
- account-wallet client - knows it's accountId, `apiUrl` and `SigningCosmWasmClient`

The commit is still WIP, yet mostly it's just the functionality of the abstract
account that is left to be implemented.

Following folders were created to manage the codebase effectively:
- `actions` – contains all actions that clients are able to perform
|  - `public` – contains all public client actions
|  - `wallet` – contains all wallet client actions
|  - `account-public` – contains all account-public client actions
|  - `account-wallet` – contains all account-wallet client actions
|  ... - base client actions
|  Some of them are grouped by module groups, such as `ans`, `account-factory`, `manager` etc.
- `utils` – contains all utility data-transformation, formatting functions
– `clients` – contains the clients implementations
|  - `decorators` – contains the decorators to inject the client data and bind those to clients
- `types` – contains types

With such a structure, we're clearly separating the business logic from the utilitary logic that
mostly concerns different data transormations.

The focus here is also to keep the functions with the most minimal API as possible, so that hardly
re-uses `ts-codegen` generated types whenever possible.

The API might not look user-friendly at the moment, but later on we will provide aliases for more
user-friendly function names.