fix: cli-react dependency
Before this commit, one could not practically use
`react` package as it would need clients to be injected,
which at the time was happening in the `cli` package.

This fix introduces new `Provider` entity that is moved
from the `cli` package and split into `provider-graz`
and `provider-cosmoskit` packages, that are now required
to be passed in the new `createConfig` method imported from
`react` package.

This also allowed to delete provider-dependant logic from `cli`
package, so now it has the same output for all providers.