feat: propagate full query result from queries
The reason being is that all of the queries rely on
`use(Abstract)ModuleQueryClient` and it is pretty straightforward to
figure out the cause of issue if it is on that level.

Full result is needed to leverage full api of `useQuery()`
(specifically, `.refetch` which is really needed atm).

Ideally, we don't want to depend on any queries internally (just like `wagmi`,
where it depends on a `config` which is either instantly retrievable or
not, depending whether it's in the provider or not, failing instantly if
not), but we can't since we depend on `graz` which does not export
actions. Nor doesn't cosmos-kit.